refactor(tests): remove [[test]] entries, move non-compiling to resurrection_ward
- Remove all [[test]] sections from webfang_core, webfang_ai, webfang_mcp,
  webfang_tui Cargo.toml (crate-local tests/ uses auto-discovery)
- Add #![cfg(feature = "ui")] to progress_tui_integration.rs (replaces
  required-features)
- Merge root tests/common/ into crates/webfang_core/tests/common/
- Move 14 non-compiling tests to resurrection_ward/ (reference old crate
  name 'webfang' instead of 'webfang_core', missing deps like assert_cmd)
