+++ crates/webfang_core/Cargo.toml
@@ -145,94 +145,7 @@ regex = { workspace = true }
 webfang_test_utils = { workspace = true }
 
 # ---------------------------------------------------------------------------
-# Wiring for the previously-orphaned root `tests/` suite.
-#
-# These `[[test]]` targets point at the root `tests/` tree (outside this
-# crate's directory) so the CLI behavioral / integration tests become
-# runnable as part of the `webfang_core` test build. They are owned by
-# `webfang_core` for dependency purposes; the binaries they exercise
-# (`webfang`) live in the `webfang_cli` crate.
-# ---------------------------------------------------------------------------
-[[test]]
-name = "behavioral"
-path = "../../tests/behavioral/main.rs"
 
-[[test]]
-name = "cli_binary"
-path = "../../tests/cli_binary_test.rs"
-
-[[test]]
-name = "sitemap_integration"
-path = "../../tests/sitemap_integration.rs"
-
-[[test]]
-name = "resume_mode_tests"
-path = "../../tests/resume_mode_tests.rs"
-
-[[test]]
-name = "sitemap_behavioral"
-path = "../../tests/sitemap_behavioral.rs"
-
-[[test]]
-name = "scraper_service_test"
-path = "../../tests/scraper_service_test.rs"
-
-[[test]]
-name = "args_test"
-path = "../../tests/args_test.rs"
-
-[[test]]
-name = "cleaning_pipelines_regression"
-path = "../../tests/cleaning_pipelines_regression.rs"
-
-[[test]]
-name = "session_pool_integration"
-path = "../../tests/session_pool_integration.rs"
-
-[[test]]
-name = "infrastructure_session_pool"
-path = "../../tests/infrastructure/session_pool.rs"
-
-[[test]]
-name = "adaptive_selector_integration"
-path = "tests/adaptive_selector_integration.rs"
-
-[[test]]
-name = "progress_tui_integration"
-path = "../../tests/progress_tui_integration.rs"
-required-features = ["ui"]
-
-[[test]]
-name = "progress_e2e_wiring"
-path = "../../tests/progress_e2e_wiring.rs"
-
-[[test]]
-name = "engine_js_strategy_timeout"
-path = "../../tests/engine_js_strategy_timeout_test.rs"
-
-[[test]]
-name = "checkpoint_store_integration"
-path = "../../tests/checkpoint_store_integration.rs"
-
-[[test]]
-name = "downloader_headers_test"
-path = "../../tests/downloader_headers_test.rs"
-
-[[test]]
-name = "http_client_integration"
-path = "../../tests/http_client_integration.rs"
-
-[[test]]
-name = "ai_error_propagation"
-path = "../../tests/ai_error_propagation.rs"
-
-[[test]]
-name = "waf_fixtures_test"
-path = "../../tests/waf_fixtures_test.rs"
-
-[[test]]
-name = "security_integration"
-path = "../../tests/security_integration.rs"
 
 # ---------------------------------------------------------------------------
 # Benchmarks (criterion). The bench files live in the root `benches/` tree
