+++ crates/webfang_mcp/Cargo.toml
@@ -50,17 +50,7 @@ insta = { workspace = true }
 scraper = { workspace = true }
 wiremock = { workspace = true }
 
-[[test]]
-name = "discover_sitemap_wiremock"
-path = "tests/discover_sitemap_wiremock.rs"
 
-[[test]]
-name = "mcp_behavioral_test"
-path = "../../tests/mcp_behavioral_test.rs"
-
-[[test]]
-name = "mcp_lifecycle_test"
-path = "../../tests/mcp_lifecycle_test.rs"
 
 [lints]
 workspace = true
