+++ crates/webfang_tui/Cargo.toml
@@ -40,9 +40,7 @@ chrono = { workspace = true }
 tempfile = { workspace = true }
 walkdir = { workspace = true }
 
-[[test]]
-name = "tui_unit_tests"
-path = "../../tests/tui_unit_tests.rs"
+
 
 [lints]
 workspace = true
