+++ resurrection_ward/vault_detector_integration.rs
